Temporary workaround for issue #13876
We've already cut everything over to using 2.0.0.2, but are now seeing
quite frequent intermittent task failures regarding temp dir creation
which is killing deployment pipelines in bamboo that we use to deploy
these changes. In the meantime we'll vendor ansible and pin to this
commit until there's a proper fix for this upstream.

This work is part of LAAS-916